backend/pyproject.toml
[tool.poetry]
name = "playbook"
version = "0.1.0"
description = ""
authors = ["Renato Magalhães <renato.n.magalhaes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-core = "^0.13.2"
llama-index-llms-google-genai = "^0.3.0"
llama-index-embeddings-google-genai = "^0.3.0"
python-dotenv = "^1.1.1"
fastapi = {extras = ["standard"], version = "^0.116.1"}
pydantic = "^2.11.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
